Schematic: Set SPI output ESD diodes to steer to VOUT
Since +5 V may not always be present, such as when the dodoprog is
connected to a target but not powered, and we need to still suppress ESD
and not present a short looking load on the SPI bus.

If there is data on the output SPI bus, there always should be a VOUT
present!